Re-added quarantine/file delete functionality from 5.x [tracker 2226]

+++ libraries/File_Scan.php
@@ -7,7 +7,7 @@
  * @package    file-scan
  * @subpackage libraries
  * @author     ClearFoundation <developer@clearfoundation.com>
- * @copyright  2006-2014 ClearFoundation
+ * @copyright  2006-2015 ClearFoundation
  * @license    http://www.gnu.org/copyleft/lgpl.html GNU Lesser General Public License version 3 or later
  * @link       http://www.clearfoundation.com/docs/developer/apps/file_scan/
  */
@@ -55,20 +55,18 @@ clearos_load_language('file_scan');
 // Classes
 //--------
 
-use \clearos\apps\base\Configuration_File as Configuration_File;
 use \clearos\apps\base\Engine as Engine;
 use \clearos\apps\base\File as File;
+use \clearos\apps\base\Configuration_File as Configuration_File;
 use \clearos\apps\base\Folder as Folder;
 use \clearos\apps\base\Shell as Shell;
-use \clearos\apps\file_scan\File_Scan as File_Scan;
 use \clearos\apps\tasks\Cron as Cron;
 
-clearos_load_library('base/Configuration_File');
 clearos_load_library('base/Engine');
 clearos_load_library('base/File');
+clearos_load_library('base/Configuration_File');
 clearos_load_library('base/Folder');
 clearos_load_library('base/Shell');
-clearos_load_library('file_scan/File_Scan');
 clearos_load_library('tasks/Cron');
 
 // Exceptions
@@ -93,7 +91,7 @@ clearos_load_library('base/Validation_Exception');
  * @package    file-scan
  * @subpackage libraries
  * @author     ClearFoundation <developer@clearfoundation.com>
- * @copyright  2006-2014 ClearFoundation
+ * @copyright  2006-2015 ClearFoundation
  * @license    http://www.gnu.org/copyleft/lgpl.html GNU Lesser General Public License version 3 or later
  * @link       http://www.clearfoundation.com/docs/developer/apps/file_scan/
  */
@@ -126,7 +124,7 @@ class File_Scan extends Engine
     const FILE_STATE = '/var/clearos/framework/tmp/avscan.state';
 
     // Locating of quarantine directory
-    const PATH_QUARANTINE = '/var/lib/quarantine';
+    const PATH_QUARANTINE = '/var/clearos/file_scan/quarantine';
 
     // Status
     const STATUS_IDLE = 0;    
@@ -184,44 +182,47 @@ class File_Scan extends Engine
 
         $file = new File(self::FILE_SCAN_FOLDERS, TRUE);
 
-        if (!$file->exists())
-            $file->create('root', 'root', '0644');
+        if (!$file->exists()) {
+            try {
+                $file->Create('root', 'root', '0644');
+            } catch (Engine_Exception $e) {
+                throw new Engine_Exception($e->get_message(), CLEAROS_ERROR);
+            }
+        }
 
-        $file->dump_contents_from_array($dirs);
-    }
-
-    /**
-     * Deletes a quarantined virus.
-     *
-     * @param string $hash MD5 hash of virus filename to delete
-     *
-     * @return void
-     * @throws Engine_Exception
-     */
-
-    public function delete_quarantined_virus($hash)
-    {
-        clearos_profile(__METHOD__, __LINE__);
-
-        $nfo = new File(self::PATH_QUARANTINE . "/$hash.nfo", TRUE);
-        $nfo->Delete();
-
-        $dat = new File(self::PATH_QUARANTINE . "/$hash.dat", TRUE);
-        $dat->Delete();
+        try {
+            $file->dump_contents_from_array($dirs);
+        } catch (Engine_Exception $e) {
+            throw new Engine_Exception($e->get_message(), CLEAROS_ERROR);
+        }
     }
 
     /**
      * Deletes a virus.
      *
-     * @param string $hash MD5 hash of virus filename to delete
+     * @param string $hash           MD5 hash of virus filename to delete
+     * @param boolean $in_quarantine boolean indicating whether file is in quarantine or not
      *
      * @throws Engine_Exception
      * @return void
      */
 
-    public function delete_virus($hash)
+    public function delete_virus($hash, $in_quarantine = FALSE)
     {
         clearos_profile(__METHOD__, __LINE__);
+
+        if ($in_quarantine) {
+            $list = $this->get_quarantined_viruses();
+            foreach ($list as $meta) {
+                if (md5($meta['name'] == $hash)) {
+                    $virus = new File(self::PATH_QUARANTINE . "/" . $meta['name'], TRUE);
+                    if ($virus->exists())
+                        $virus->delete();
+                    break;
+                }
+            }
+            return;
+        }
 
         if (!file_exists(self::FILE_STATE))
             throw new Engine_Exception(lang('file_scan_state_error'), CLEAROS_ERROR);
@@ -240,8 +241,13 @@ class File_Scan extends Engine
         if (!isset($this->state['virus'][$hash]))
             throw new Engine_Exception(lang('base_file_not_found'), CLEAROS_ERROR);
 
-        $virus = new File($this->state['virus'][$hash]['filename'], TRUE);
-        $virus->delete();
+        try {
+            $virus = new File($this->state['virus'][$hash]['filename'], TRUE);
+            if ($virus->exists())
+                $virus->delete();
+        } catch (Engine_Exception $e) {
+            throw new Engine_Exception($e->get_message(), CLEAROS_ERROR);
+        }
 
         // Update state file, delete virus
         unset($this->state['virus'][$hash]);
@@ -486,6 +492,27 @@ class File_Scan extends Engine
     }
 
     /**
+     * Set the quarantine action
+     *
+     * @param boolean $quarantine quarantine
+     *
+     * @return void
+     * @throws Validation_Exception
+     */
+
+    function set_quarantine($quarantine)
+    {
+        clearos_profile(__METHOD__, __LINE__);
+
+        Validation_Exception::is_valid($this->validate_quarantine($quarantine));
+
+        if ($quarantine === 'on' || $quarantine == 1 || $quarantine == TRUE)
+            $this->_set_parameter('quarantine', 1);
+        else
+            $this->_set_parameter('quarantine', 0);
+    }
+
+    /**
      * Set the notify on virus setting.
      *
      * @param boolean $notify notify
@@ -557,26 +584,18 @@ class File_Scan extends Engine
     {
         clearos_profile(__METHOD__, __LINE__);
 
+        $files = array();
         try {
-            $dir = new Folder(File_Scan::PATH_QUARANTINE, TRUE);
-            $files = $dir->get_listing();
+            $dir = new Folder(self::PATH_QUARANTINE, TRUE);
+            $files = $dir->get_listing(TRUE, TRUE);
         } catch (Folder_Not_Found_Exception $e) {
             return array();
+        } catch (Engine_Exception $e) {
+            throw new Engine_Exception($e->get_message(), CLEAROS_ERROR);
         }
 
-        $viruses = array();
-
-        foreach ($files as $file) {
-            if (stristr($file, '.nfo') === FALSE)
-                continue;
-
-            $nfo = new File(File_Scan::PATH_QUARANTINE . "/$file", TRUE);
-            $buffer = unserialize($nfo->get_contents());
-
-            $viruses[md5($buffer['filename'])] = $buffer;
-        }
-
-        return $viruses;
+        
+        return $files;
     }
 
     /**
@@ -597,8 +616,12 @@ class File_Scan extends Engine
 
         if (!$cron->exists_configlet('app-file-scan')) return array('*', '*', '*');
 
-        list($minute, $hour, $day_of_month, $month, $day_of_week) 
-            = explode(' ', $cron->get_configlet('app-file-scan'), 5);
+        try {
+            list($minute, $hour, $day_of_month, $month, $day_of_week) 
+                = explode(' ', $cron->get_configlet('app-file-scan'), 5);
+        } catch (Engine_Exception $e) {
+            throw new Engine_Exception($e->get_message(), CLEAROS_ERROR);
+        }
 
         $schedule['hour'] = $hour;
         $schedule['day_of_month'] = $day_of_month;
@@ -646,6 +669,10 @@ class File_Scan extends Engine
     {
         clearos_profile(__METHOD__, __LINE__);
 
+        $dir = new Folder(self::PATH_QUARANTINE, TRUE);
+        if (!$dir->exists())
+            $dir->create('root', 'root', 600);
+
         if (!file_exists(File_Scan::FILE_STATE))
             throw new Engine_Exception(lang('file_scan_state_error'), CLEAROS_ERROR);
 
@@ -663,12 +690,13 @@ class File_Scan extends Engine
         if (!isset($this->state['virus'][$hash]))
             throw new Engine_Exception(lang('base_file_not_found'), CLEAROS_ERROR);
 
-        $virus = new File($this->state['virus'][$hash]['filename'], TRUE);
-        $virus->move_to(File_Scan::PATH_QUARANTINE . "/$hash.dat");
-        $virus = new File(File_Scan::PATH_QUARANTINE . "/$hash.nfo");
+        try {
+            $virus = new File($this->state['virus'][$hash]['filename'], TRUE);
+            $virus->move_to(self::PATH_QUARANTINE . '/' . basename($this->state['virus'][$hash]['filename']) . '.' . date('ymd_His'));
 
-        $virus->create('webconfig', 'webconfig', '0640');
-        $virus->add_lines(serialize($this->state['virus'][$hash]));
+        } catch (Engine_Exception $e) {
+            throw new Engine_Exception($e->get_message(), CLEAROS_ERROR);
+        }
 
         // Update state file, delete virus
         unset($this->state['virus'][$hash]);
@@ -702,7 +730,11 @@ class File_Scan extends Engine
 
         $file = new File(self::FILE_SCAN_FOLDERS, TRUE);
 
-        $file->dump_contents_from_array($dirs);
+        try {
+            $file->dump_contents_from_array($dirs);
+        } catch (Engine_Exception $e) {
+            throw new Engine_Exception($e->get_message(), CLEAROS_ERROR);
+        }
     }
 
     /**
@@ -718,8 +750,12 @@ class File_Scan extends Engine
 
         $cron = new Cron();
 
-        if ($cron->exists_configlet('app-file-scan'))
-            $cron->delete_configlet('app-file-scan');
+        try {
+            if ($cron->exists_configlet('app-file-scan'))
+                $cron->delete_configlet('app-file-scan');
+        } catch (Engine_Exception $e) {
+            throw new Engine_Exception($e->get_message(), CLEAROS_ERROR);
+        }
     }
 
     /**
@@ -731,10 +767,13 @@ class File_Scan extends Engine
     public function delete_state()
     {
         clearos_profile(__METHOD__, __LINE__);
-
-        $file = new File(self::FILE_STATE, FALSE);
-        if ($file->exists())
-            $file->delete();
+        try {
+            $file = new File(self::FILE_STATE, FALSE);
+            if ($file->exists())
+                $file->delete();
+        } catch (Exception $e) {
+            throw new Engine_Exception(clearos_exception_message($e));
+        }
     }
 
     /**
@@ -756,7 +795,6 @@ class File_Scan extends Engine
         $this->state['error'] = array();
         $this->state['virus'] = array();
         $this->state['timestamp'] = 0;
-
         unset($this->state['stats']);
     }
 
@@ -773,12 +811,16 @@ class File_Scan extends Engine
     {
         clearos_profile(__METHOD__, __LINE__);
 
-        $nfo = new File(File_Scan::PATH_QUARANTINE . "/$hash.nfo", TRUE);
-        $virus = unserialize($nfo->get_contents());
+        try {
+            $nfo = new File(self::PATH_QUARANTINE . "/$hash.nfo", TRUE);
+            $virus = unserialize($nfo->get_contents());
 
-        $dat = new File(File_Scan::PATH_QUARANTINE . "/$hash.dat", TRUE);
-        $dat->move_to($virus['filename']);
-        $nfo->delete();
+            $dat = new File(self::PATH_QUARANTINE . "/$hash.dat", TRUE);
+            $dat->move_to($virus['filename']);
+            $nfo->delete();
+        } catch (Engine_Exception $e) {
+            throw new Engine_Exception($e->get_message(), CLEAROS_ERROR);
+        }
     }
 
     /**
@@ -877,11 +919,16 @@ class File_Scan extends Engine
         clearos_profile(__METHOD__, __LINE__);
 
         if ($this->is_scan_running())
-            throw new Engine_Exception(lang('file_scan_file_scan_already_running'));
+            throw new Engine_Exception(lang('file_scan_scanner_already_running'));
 
         $dirs = $this->get_directories();
         if (empty($dirs))
             throw new Engine_Exception(lang('file_scan_no_folders_selected'));
+
+        // Make sure path exists
+        $dir = new Folder(self::PATH_QUARANTINE, TRUE);
+        if (!$dir->exists())
+            $dir->create('root', 'root', 600);
 
         $options = array();
         $options['background'] = TRUE;
@@ -951,6 +998,27 @@ class File_Scan extends Engine
             return FALSE;
 
         return TRUE;
+    }
+
+    /**
+     * Get the quarantine setting
+     *
+     * @return Boolean
+     */
+
+    function get_quarantine()
+    {
+        clearos_profile(__METHOD__, __LINE__);
+
+        if (!$this->is_loaded)
+            $this->_load_config();
+
+        $quarantine = $this->config['quarantine'];
+
+        if ($quarantine == NULL || !$quarantine)
+            return FALSE;
+        else
+            return TRUE;
     }
 
     /**
@@ -1030,7 +1098,11 @@ class File_Scan extends Engine
 
         $configfile = new Configuration_File(self::FILE_CONFIG);
 
-        $this->config = $configfile->load();
+        try {
+            $this->config = $configfile->load();
+        } catch (Exception $e) {
+            throw new Engine_Exception(clearos_exception_message($e), CLEAROS_ERROR);
+        }
 
         $this->is_loaded = TRUE;
     }
@@ -1049,15 +1121,19 @@ class File_Scan extends Engine
     {
         clearos_profile(__METHOD__, __LINE__);
 
-        $file = new File(self::FILE_CONFIG, TRUE);
+        try {
+            $file = new File(self::FILE_CONFIG, TRUE);
 
-        if (!$file->exists())
-            $file->create('webconfig', 'webconfig', '0644');
+            if (!$file->exists())
+                $file->create('webconfig', 'webconfig', '0644');
 
-        $match = $file->replace_lines("/^$key\s*=\s*/", "$key=$value\n");
+            $match = $file->replace_lines("/^$key\s*=\s*/", "$key=$value\n");
 
-        if (!$match)
-            $file->add_lines("$key=$value\n");
+            if (!$match)
+                $file->add_lines("$key=$value\n");
+        } catch (Exception $e) {
+            throw new Engine_Exception(clearos_exception_message($e), CLEAROS_ERROR);
+        }
 
         $this->is_loaded = FALSE;
     }
@@ -1104,6 +1180,18 @@ class File_Scan extends Engine
         clearos_profile(__METHOD__, __LINE__);
     }
 
+    /**
+     * Validation routine for quarantine.
+     *
+     * @param boolean $quarantine quarantine
+     *
+     * @return mixed void if quarantine on error is valid, errmsg otherwise
+     */
+
+    public function validate_quarantine($quarantine)
+    {
+        clearos_profile(__METHOD__, __LINE__);
+    }
 }
 
 // vi: ts=4
